Move Trezor class to device.trezor and place mock_trezorlib in tests.conftest
