+++ nucypher/cli/main.py
@@ -26,8 +26,8 @@ from nucypher.cli import status
 from nucypher.cli.characters import moe, ursula, alice, bob, enrico, felix
 from nucypher.cli.config import nucypher_click_config, NucypherClickConfig
 from nucypher.cli.device.actions import DEVICE_ACTIONS
-from nucypher.cli.hardware.backends import Trezor
 from nucypher.cli.painting import echo_version
+from nucypher.device.trezor import Trezor
 from nucypher.network.middleware import RestMiddleware
 from nucypher.utilities.logging import GlobalConsoleLogger, getJsonFileObserver, SimpleObserver, logToSentry
 from nucypher.utilities.sandbox.middleware import MockRestMiddleware
